+++ runtime/vm/raw_object.h
@@ -2726,7 +2726,9 @@ class ReceivePortLayout : public InstanceLayout {
   VISIT_FROM(ObjectPtr, send_port_)
   SendPortPtr send_port_;
   InstancePtr handler_;
-  VISIT_TO(ObjectPtr, handler_)
+  StringPtr debug_name_;
+  StackTracePtr allocation_location_;
+  VISIT_TO(ObjectPtr, allocation_location_)
 };
 
 class TransferableTypedDataLayout : public InstanceLayout {